+++ examples/basic/README.md
@@ -1,0 +1,8 @@
+# Basic example
+
+A basic example that shows how to provision a Key Protect instance, Key Ring, and Keys.
+
+The following resources are provisioned by this example:
+- A new resource group, if an existing one is not passed in.
+- A Key Protect instance in the resource group and region provided.
+- A new Key Ring with multiple new Keys in the newly provisioned Key Protect instance.
