+++ examples/basic/outputs.tf
