+++ examples/basic/version.tf
@@ -2,11 +2,11 @@ terraform {
   required_version = ">= 1.0.0, <1.7.0"
 
   # Ensure that there is always 1 example locked into the lowest provider version of the range defined in the main
-  # module's version.tf (basic example), and 1 example that will always use the latest provider version (this example).
+  # module's version.tf (this example), and 1 example that will always use the latest provider version (existing resources example).
   required_providers {
     ibm = {
       source  = "IBM-Cloud/ibm"
-      version = ">= 1.58.0"
+      version = "1.58.0"
     }
   }
 }
