--- examples/default/README.md
@@ -1,8 +1,0 @@
-# End to end example with default values
-
-An end-to-end example that uses the module's default variable values.
-This example will:
-- Create a new resource group (if existing one is not passed in).
-- Create a new Key Protect instance in the region provided.
-- Enable metrics on the Key Protect instance.
-- Create new key rings and keys on the Key Protect instance.
