+++ examples/existing-resources/variables.tf
@@ -6,20 +6,8 @@ variable "ibmcloud_api_key" {
 
 variable "region" {
   type        = string
-  description = "Region to provision all resources created by this example"
+  description = "Region in which existing KMS instance exists"
   default     = "us-south"
-}
-
-variable "prefix" {
-  type        = string
-  description = "Prefix to append to all resources created by this example"
-  default     = "kp-all-inc-existing"
-}
-
-variable "resource_group" {
-  type        = string
-  description = "An existing resource group name to use for this example, if unset a new resource group will be created"
-  default     = null
 }
 
 variable "existing_kms_instance_guid" {
