--- examples/slz/README.md
@@ -1,8 +1,0 @@
-# Example with SLZ default values
-
-An end-to-end example that uses SLZ's default Key Protect variable values.
-This example will:
-- Create a new resource group (if existing one is not passed in).
-- Create a new Key Protect instance in the region provided.
-- Enable metrics on the Key Protect instance.
-- Create a new key ring and keys on the Key Protect instance.
