--- examples/slz/providers.tf
@@ -1,4 +1,0 @@
-provider "ibm" {
-  ibmcloud_api_key = var.ibmcloud_api_key
-  region           = var.region
-}
