+++ version.tf
@@ -1,7 +1,7 @@
 terraform {
   required_version = ">= 1.0.0, <1.7.0"
 
-  # Use "greater than or equal to" range in modules
+  # Each required provider's version should be a flexible range to future proof the module's usage with upcoming minor and patch versions.
   required_providers {
     # ignore linter error - include providers required by all child modules
     # tflint-ignore: terraform_unused_required_providers
